In MainViewModel and other relevant files:   - Changed type of the BoardSize property from byte to int
In InputPanelUserControl
  - Added error message TextBlock with binding and visibility converter
  - Implemented GroupBox for input with labels, TextBox, ComboBox, and Slider
  - Bound UI elements to ViewModel properties for dynamic data handling
  - Styled UI elements using StaticResource references

+++ NQueen.Console/Utils/DispatchUtils.cs
@@ -47,10 +47,10 @@ public static class DispatchUtils
         return board;
     }
 
-    public static (bool isValid, byte boardSize) CheckBoardSize(
+    public static (bool isValid, int boardSize) CheckBoardSize(
         string value, SolutionMode solutionMode)
     {
-        if (byte.TryParse(value, out byte size) == false)
+        if (int.TryParse(value, out int size) == false)
         {
             HelpCommands.ShowExitError(CommandConst.InvalidBoardSize);
             return (false, 0);
