+++ NQueen.Kernel/Utilities/BoardSettings.cs
@@ -2,14 +2,15 @@
 
 public static class BoardSettings
 {
+    // Todo: Remove or change this property
     public const byte ByteMaxValue = byte.MaxValue;
-    public const byte DefaultBoardSize = 8;
-    public const byte MinBoardSize = 1;
+    public const int DefaultBoardSize = 8;
+    public const int MinBoardSize = 1;
 
-    public const byte MaxBoardSizeInSingleSolution = 37;
-    public const byte MaxBoardSizeInUniqueSolutions = 17;
-    public const byte MaxBoardSizeInAllSolutions = 17;
+    public const int MaxBoardSizeInSingleSolution = 37;
+    public const int MaxBoardSizeInUniqueSolutions = 17;
+    public const int MaxBoardSizeInAllSolutions = 17;
 
-    public const byte SmallBoardSizeInUniqueSolutions = 10;
-    public const byte MediumBoardSizeInUniqueSolutions = 15;
+    public const int SmallBoardSizeInUniqueSolutions = 10;
+    public const int MediumBoardSizeInUniqueSolutions = 15;
 }
